Removed AdvancedDiv and Moved it to Generics
